consolidando dialogs

+++ src/app/modules/authentication/user/list/user-list.component.ts
@@ -1,3 +1,4 @@
+import { DialogService } from './../../../../../@fuse/services/dialogs/dialog.service';
 import { Router } from '@angular/router';
 import { ApiService } from './../../../../api/services/api.service';
 import { FuseAlertType } from './../../../../../@fuse/components/alert/alert.types';
@@ -26,11 +27,13 @@ export class UserListComponent implements OnInit {
     displayedColumns: string[] = ['name', 'email', 'profile', 'actions'];
 
     rota = 'authentication/user';
-
+    widthScreen = window.screen.width;
+    
     constructor(
         private readonly _api: ApiService,
         private readonly router: Router,
-        private readonly dc: ChangeDetectorRef
+        private readonly dc: ChangeDetectorRef,
+        private readonly dialogService: DialogService
     ) {}
 
     ngOnInit(): void {
@@ -52,22 +55,39 @@ export class UserListComponent implements OnInit {
     }
 
     deleteRow(row) {
-        this._api.deleteUser(row.id).subscribe(
-            (res) => {
-                this.loadingRows();
-                this.dc.detectChanges();
-                
-                this.alert = {
-                    type: 'success',
-                    message: 'Registro removido com sucesso',
-                };
-            },
-            (err) => {
-                this.alert = {
-                    type: 'error',
-                    message: 'Ocorreu um erro, tente novamente!',
-                };
-            }
-        );
+        let width = null;
+
+        if (+this.widthScreen < 450 ) {
+            width = '265px;'
+        }
+
+        this.dialogService
+            .confirm(
+                'Com essa ação, o usuário será removido, permanentemente!',
+                'Remover usuário',
+                true,
+                { btnOkText: 'Remover', width }
+            )
+            .then((res) => {
+                if (res) {
+                    this._api.deleteUser(row.id).subscribe(
+                        (res) => {
+                            this.loadingRows();
+                            this.dc.detectChanges();
+
+                            this.alert = {
+                                type: 'success',
+                                message: 'Registro removido com sucesso',
+                            };
+                        },
+                        (err) => {
+                            this.alert = {
+                                type: 'error',
+                                message: 'Ocorreu um erro, tente novamente!',
+                            };
+                        }
+                    );
+                }
+            });
     }
 }
